Struct aws_sdk_globalaccelerator::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for GlobalAccelerator_V20180706.

This client allows ergonomic access to a GlobalAccelerator_V20180706-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

Creates a client with the given service configuration.

pub fn conf(&self) -> &Config

Returns the client’s configuration.

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn add_custom_routing_endpoints(&self) -> AddCustomRoutingEndpoints<C, M, R>

Constructs a fluent builder for the AddCustomRoutingEndpoints operation.

See AddCustomRoutingEndpoints for more information about the operation and its arguments.

pub fn advertise_byoip_cidr(&self) -> AdvertiseByoipCidr<C, M, R>

Constructs a fluent builder for the AdvertiseByoipCidr operation.

See AdvertiseByoipCidr for more information about the operation and its arguments.

pub fn allow_custom_routing_traffic(&self) -> AllowCustomRoutingTraffic<C, M, R>

Constructs a fluent builder for the AllowCustomRoutingTraffic operation.

See AllowCustomRoutingTraffic for more information about the operation and its arguments.

pub fn create_accelerator(&self) -> CreateAccelerator<C, M, R>

Constructs a fluent builder for the CreateAccelerator operation.

See CreateAccelerator for more information about the operation and its arguments.

pub fn create_custom_routing_accelerator(
&self
) -> CreateCustomRoutingAccelerator<C, M, R>

Constructs a fluent builder for the CreateCustomRoutingAccelerator operation.

See CreateCustomRoutingAccelerator for more information about the operation and its arguments.

pub fn create_custom_routing_endpoint_group(
&self
) -> CreateCustomRoutingEndpointGroup<C, M, R>

Constructs a fluent builder for the CreateCustomRoutingEndpointGroup operation.

See CreateCustomRoutingEndpointGroup for more information about the operation and its arguments.

pub fn create_custom_routing_listener(
&self
) -> CreateCustomRoutingListener<C, M, R>

Constructs a fluent builder for the CreateCustomRoutingListener operation.

See CreateCustomRoutingListener for more information about the operation and its arguments.

pub fn create_endpoint_group(&self) -> CreateEndpointGroup<C, M, R>

Constructs a fluent builder for the CreateEndpointGroup operation.

See CreateEndpointGroup for more information about the operation and its arguments.

pub fn create_listener(&self) -> CreateListener<C, M, R>

Constructs a fluent builder for the CreateListener operation.

See CreateListener for more information about the operation and its arguments.

pub fn delete_accelerator(&self) -> DeleteAccelerator<C, M, R>

Constructs a fluent builder for the DeleteAccelerator operation.

See DeleteAccelerator for more information about the operation and its arguments.

pub fn delete_custom_routing_accelerator(
&self
) -> DeleteCustomRoutingAccelerator<C, M, R>

Constructs a fluent builder for the DeleteCustomRoutingAccelerator operation.

See DeleteCustomRoutingAccelerator for more information about the operation and its arguments.

pub fn delete_custom_routing_endpoint_group(
&self
) -> DeleteCustomRoutingEndpointGroup<C, M, R>

Constructs a fluent builder for the DeleteCustomRoutingEndpointGroup operation.

See DeleteCustomRoutingEndpointGroup for more information about the operation and its arguments.

pub fn delete_custom_routing_listener(
&self
) -> DeleteCustomRoutingListener<C, M, R>

Constructs a fluent builder for the DeleteCustomRoutingListener operation.

See DeleteCustomRoutingListener for more information about the operation and its arguments.

pub fn delete_endpoint_group(&self) -> DeleteEndpointGroup<C, M, R>

Constructs a fluent builder for the DeleteEndpointGroup operation.

See DeleteEndpointGroup for more information about the operation and its arguments.

pub fn delete_listener(&self) -> DeleteListener<C, M, R>

Constructs a fluent builder for the DeleteListener operation.

See DeleteListener for more information about the operation and its arguments.

pub fn deny_custom_routing_traffic(&self) -> DenyCustomRoutingTraffic<C, M, R>

Constructs a fluent builder for the DenyCustomRoutingTraffic operation.

See DenyCustomRoutingTraffic for more information about the operation and its arguments.

pub fn deprovision_byoip_cidr(&self) -> DeprovisionByoipCidr<C, M, R>

Constructs a fluent builder for the DeprovisionByoipCidr operation.

See DeprovisionByoipCidr for more information about the operation and its arguments.

pub fn describe_accelerator(&self) -> DescribeAccelerator<C, M, R>

Constructs a fluent builder for the DescribeAccelerator operation.

See DescribeAccelerator for more information about the operation and its arguments.

pub fn describe_accelerator_attributes(
&self
) -> DescribeAcceleratorAttributes<C, M, R>

Constructs a fluent builder for the DescribeAcceleratorAttributes operation.

See DescribeAcceleratorAttributes for more information about the operation and its arguments.

pub fn describe_custom_routing_accelerator(
&self
) -> DescribeCustomRoutingAccelerator<C, M, R>

Constructs a fluent builder for the DescribeCustomRoutingAccelerator operation.

See DescribeCustomRoutingAccelerator for more information about the operation and its arguments.

pub fn describe_custom_routing_accelerator_attributes(
&self
) -> DescribeCustomRoutingAcceleratorAttributes<C, M, R>

Constructs a fluent builder for the DescribeCustomRoutingAcceleratorAttributes operation.

See DescribeCustomRoutingAcceleratorAttributes for more information about the operation and its arguments.

pub fn describe_custom_routing_endpoint_group(
&self
) -> DescribeCustomRoutingEndpointGroup<C, M, R>

Constructs a fluent builder for the DescribeCustomRoutingEndpointGroup operation.

See DescribeCustomRoutingEndpointGroup for more information about the operation and its arguments.

pub fn describe_custom_routing_listener(
&self
) -> DescribeCustomRoutingListener<C, M, R>

Constructs a fluent builder for the DescribeCustomRoutingListener operation.

See DescribeCustomRoutingListener for more information about the operation and its arguments.

pub fn describe_endpoint_group(&self) -> DescribeEndpointGroup<C, M, R>

Constructs a fluent builder for the DescribeEndpointGroup operation.

See DescribeEndpointGroup for more information about the operation and its arguments.

pub fn describe_listener(&self) -> DescribeListener<C, M, R>

Constructs a fluent builder for the DescribeListener operation.

See DescribeListener for more information about the operation and its arguments.

pub fn list_accelerators(&self) -> ListAccelerators<C, M, R>

Constructs a fluent builder for the ListAccelerators operation.

See ListAccelerators for more information about the operation and its arguments.

pub fn list_byoip_cidrs(&self) -> ListByoipCidrs<C, M, R>

Constructs a fluent builder for the ListByoipCidrs operation.

See ListByoipCidrs for more information about the operation and its arguments.

pub fn list_custom_routing_accelerators(
&self
) -> ListCustomRoutingAccelerators<C, M, R>

Constructs a fluent builder for the ListCustomRoutingAccelerators operation.

See ListCustomRoutingAccelerators for more information about the operation and its arguments.

pub fn list_custom_routing_endpoint_groups(
&self
) -> ListCustomRoutingEndpointGroups<C, M, R>

Constructs a fluent builder for the ListCustomRoutingEndpointGroups operation.

See ListCustomRoutingEndpointGroups for more information about the operation and its arguments.

pub fn list_custom_routing_listeners(
&self
) -> ListCustomRoutingListeners<C, M, R>

Constructs a fluent builder for the ListCustomRoutingListeners operation.

See ListCustomRoutingListeners for more information about the operation and its arguments.

pub fn list_custom_routing_port_mappings(
&self
) -> ListCustomRoutingPortMappings<C, M, R>

Constructs a fluent builder for the ListCustomRoutingPortMappings operation.

See ListCustomRoutingPortMappings for more information about the operation and its arguments.

pub fn list_custom_routing_port_mappings_by_destination(
&self
) -> ListCustomRoutingPortMappingsByDestination<C, M, R>

Constructs a fluent builder for the ListCustomRoutingPortMappingsByDestination operation.

See ListCustomRoutingPortMappingsByDestination for more information about the operation and its arguments.

pub fn list_endpoint_groups(&self) -> ListEndpointGroups<C, M, R>

Constructs a fluent builder for the ListEndpointGroups operation.

See ListEndpointGroups for more information about the operation and its arguments.

pub fn list_listeners(&self) -> ListListeners<C, M, R>

Constructs a fluent builder for the ListListeners operation.

See ListListeners for more information about the operation and its arguments.

pub fn list_tags_for_resource(&self) -> ListTagsForResource<C, M, R>

Constructs a fluent builder for the ListTagsForResource operation.

See ListTagsForResource for more information about the operation and its arguments.

pub fn provision_byoip_cidr(&self) -> ProvisionByoipCidr<C, M, R>

Constructs a fluent builder for the ProvisionByoipCidr operation.

See ProvisionByoipCidr for more information about the operation and its arguments.

pub fn remove_custom_routing_endpoints(
&self
) -> RemoveCustomRoutingEndpoints<C, M, R>

Constructs a fluent builder for the RemoveCustomRoutingEndpoints operation.

See RemoveCustomRoutingEndpoints for more information about the operation and its arguments.

pub fn tag_resource(&self) -> TagResource<C, M, R>

Constructs a fluent builder for the TagResource operation.

See TagResource for more information about the operation and its arguments.

pub fn untag_resource(&self) -> UntagResource<C, M, R>

Constructs a fluent builder for the UntagResource operation.

See UntagResource for more information about the operation and its arguments.

pub fn update_accelerator(&self) -> UpdateAccelerator<C, M, R>

Constructs a fluent builder for the UpdateAccelerator operation.

See UpdateAccelerator for more information about the operation and its arguments.

pub fn update_accelerator_attributes(
&self
) -> UpdateAcceleratorAttributes<C, M, R>

Constructs a fluent builder for the UpdateAcceleratorAttributes operation.

See UpdateAcceleratorAttributes for more information about the operation and its arguments.

pub fn update_custom_routing_accelerator(
&self
) -> UpdateCustomRoutingAccelerator<C, M, R>

Constructs a fluent builder for the UpdateCustomRoutingAccelerator operation.

See UpdateCustomRoutingAccelerator for more information about the operation and its arguments.

pub fn update_custom_routing_accelerator_attributes(
&self
) -> UpdateCustomRoutingAcceleratorAttributes<C, M, R>

Constructs a fluent builder for the UpdateCustomRoutingAcceleratorAttributes operation.

See UpdateCustomRoutingAcceleratorAttributes for more information about the operation and its arguments.

pub fn update_custom_routing_listener(
&self
) -> UpdateCustomRoutingListener<C, M, R>

Constructs a fluent builder for the UpdateCustomRoutingListener operation.

See UpdateCustomRoutingListener for more information about the operation and its arguments.

pub fn update_endpoint_group(&self) -> UpdateEndpointGroup<C, M, R>

Constructs a fluent builder for the UpdateEndpointGroup operation.

See UpdateEndpointGroup for more information about the operation and its arguments.

pub fn update_listener(&self) -> UpdateListener<C, M, R>

Constructs a fluent builder for the UpdateListener operation.

See UpdateListener for more information about the operation and its arguments.

pub fn withdraw_byoip_cidr(&self) -> WithdrawByoipCidr<C, M, R>

Constructs a fluent builder for the WithdrawByoipCidr operation.

See WithdrawByoipCidr for more information about the operation and its arguments.

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

Creates a client with the given service config and connector override.

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

Creates a new client from a shared config.

pub fn from_conf(conf: Config) -> Self

Creates a new client from the service Config.