Module aws_sdk_globalaccelerator::error

Modules

accelerator_not_disabled_exception

See AcceleratorNotDisabledException

accelerator_not_found_exception

See AcceleratorNotFoundException

access_denied_exception

See AccessDeniedException

associated_endpoint_group_found_exception

See AssociatedEndpointGroupFoundException

associated_listener_found_exception

See AssociatedListenerFoundException

byoip_cidr_not_found_exception

See ByoipCidrNotFoundException

conflict_exception

See ConflictException

endpoint_already_exists_exception

See EndpointAlreadyExistsException

endpoint_group_already_exists_exception

See EndpointGroupAlreadyExistsException

endpoint_group_not_found_exception

See EndpointGroupNotFoundException

endpoint_not_found_exception

See EndpointNotFoundException

incorrect_cidr_state_exception

See IncorrectCidrStateException

internal_service_error_exception

See InternalServiceErrorException

invalid_argument_exception

See InvalidArgumentException

invalid_next_token_exception

See InvalidNextTokenException

invalid_port_range_exception

See InvalidPortRangeException

limit_exceeded_exception

See LimitExceededException

listener_not_found_exception

See ListenerNotFoundException

Structs

AcceleratorNotDisabledException

The accelerator that you specified could not be disabled.

AcceleratorNotFoundException

The accelerator that you specified doesn't exist.

AccessDeniedException

You don't have access permission.

AddCustomRoutingEndpointsError

AdvertiseByoipCidrError

AllowCustomRoutingTrafficError

AssociatedEndpointGroupFoundException

The listener that you specified has an endpoint group associated with it. You must remove all dependent resources

AssociatedListenerFoundException

The accelerator that you specified has a listener associated with it. You must remove all dependent resources from an

ByoipCidrNotFoundException

The CIDR that you specified was not found or is incorrect.

ConflictException

You can't use both of those options.

CreateAcceleratorError

CreateCustomRoutingAcceleratorError

CreateCustomRoutingEndpointGroupError

CreateCustomRoutingListenerError

CreateEndpointGroupError

CreateListenerError

DeleteAcceleratorError

DeleteCustomRoutingAcceleratorError

DeleteCustomRoutingEndpointGroupError

DeleteCustomRoutingListenerError

DeleteEndpointGroupError

DeleteListenerError

DenyCustomRoutingTrafficError

DeprovisionByoipCidrError

DescribeAcceleratorAttributesError

DescribeAcceleratorError

DescribeCustomRoutingAcceleratorAttributesError

DescribeCustomRoutingAcceleratorError

DescribeCustomRoutingEndpointGroupError

DescribeCustomRoutingListenerError

DescribeEndpointGroupError

DescribeListenerError

EndpointAlreadyExistsException

The endpoint that you specified doesn't exist.

EndpointGroupAlreadyExistsException

The endpoint group that you specified already exists.

EndpointGroupNotFoundException

The endpoint group that you specified doesn't exist.

EndpointNotFoundException

The endpoint that you specified doesn't exist.

IncorrectCidrStateException

The CIDR that you specified is not valid for this action. For example, the state of the CIDR might be

InternalServiceErrorException

There was an internal error for AWS Global Accelerator.

InvalidArgumentException

An argument that you specified is invalid.

InvalidNextTokenException

There isn't another item to return.

InvalidPortRangeException

The port numbers that you specified are not valid numbers or are not unique for this accelerator.

LimitExceededException

Processing your request would cause you to exceed an AWS Global Accelerator limit.

ListAcceleratorsError

ListByoipCidrsError

ListCustomRoutingAcceleratorsError

ListCustomRoutingEndpointGroupsError

ListCustomRoutingListenersError

ListCustomRoutingPortMappingsByDestinationError

ListCustomRoutingPortMappingsError

ListEndpointGroupsError

ListListenersError

ListTagsForResourceError

ListenerNotFoundException

The listener that you specified doesn't exist.

ProvisionByoipCidrError

RemoveCustomRoutingEndpointsError

TagResourceError

UntagResourceError

UpdateAcceleratorAttributesError

UpdateAcceleratorError

UpdateCustomRoutingAcceleratorAttributesError

UpdateCustomRoutingAcceleratorError

UpdateCustomRoutingListenerError

UpdateEndpointGroupError

UpdateListenerError

WithdrawByoipCidrError

Enums

AddCustomRoutingEndpointsErrorKind

AdvertiseByoipCidrErrorKind

AllowCustomRoutingTrafficErrorKind

CreateAcceleratorErrorKind

CreateCustomRoutingAcceleratorErrorKind

CreateCustomRoutingEndpointGroupErrorKind

CreateCustomRoutingListenerErrorKind

CreateEndpointGroupErrorKind

CreateListenerErrorKind

DeleteAcceleratorErrorKind

DeleteCustomRoutingAcceleratorErrorKind

DeleteCustomRoutingEndpointGroupErrorKind

DeleteCustomRoutingListenerErrorKind

DeleteEndpointGroupErrorKind

DeleteListenerErrorKind

DenyCustomRoutingTrafficErrorKind

DeprovisionByoipCidrErrorKind

DescribeAcceleratorAttributesErrorKind

DescribeAcceleratorErrorKind

DescribeCustomRoutingAcceleratorAttributesErrorKind

DescribeCustomRoutingAcceleratorErrorKind

DescribeCustomRoutingEndpointGroupErrorKind

DescribeCustomRoutingListenerErrorKind

DescribeEndpointGroupErrorKind

DescribeListenerErrorKind

ListAcceleratorsErrorKind

ListByoipCidrsErrorKind

ListCustomRoutingAcceleratorsErrorKind

ListCustomRoutingEndpointGroupsErrorKind

ListCustomRoutingListenersErrorKind

ListCustomRoutingPortMappingsByDestinationErrorKind

ListCustomRoutingPortMappingsErrorKind

ListEndpointGroupsErrorKind

ListListenersErrorKind

ListTagsForResourceErrorKind

ProvisionByoipCidrErrorKind

RemoveCustomRoutingEndpointsErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateAcceleratorAttributesErrorKind

UpdateAcceleratorErrorKind

UpdateCustomRoutingAcceleratorAttributesErrorKind

UpdateCustomRoutingAcceleratorErrorKind

UpdateCustomRoutingListenerErrorKind

UpdateEndpointGroupErrorKind

UpdateListenerErrorKind

WithdrawByoipCidrErrorKind