Module aws_sdk_globalaccelerator::model

Modules

accelerator

See Accelerator

accelerator_attributes

See AcceleratorAttributes

byoip_cidr

See ByoipCidr

byoip_cidr_event

See ByoipCidrEvent

cidr_authorization_context

See CidrAuthorizationContext

custom_routing_accelerator

See CustomRoutingAccelerator

custom_routing_accelerator_attributes

See CustomRoutingAcceleratorAttributes

custom_routing_destination_configuration

See CustomRoutingDestinationConfiguration

custom_routing_destination_description

See CustomRoutingDestinationDescription

custom_routing_endpoint_configuration

See CustomRoutingEndpointConfiguration

custom_routing_endpoint_description

See CustomRoutingEndpointDescription

custom_routing_endpoint_group

See CustomRoutingEndpointGroup

custom_routing_listener

See CustomRoutingListener

destination_port_mapping

See DestinationPortMapping

endpoint_configuration

See EndpointConfiguration

endpoint_description

See EndpointDescription

endpoint_group

See EndpointGroup

ip_set

See IpSet

listener

See Listener

port_mapping

See PortMapping

port_override

See PortOverride

port_range

See PortRange

socket_address

See SocketAddress

tag

See Tag

Structs

Accelerator

An accelerator is a complex type that includes one or more listeners that process inbound connections and then direct

AcceleratorAttributes

Attributes of an accelerator.

ByoipCidr

Information about an IP address range that is provisioned for use with your AWS resources through

ByoipCidrEvent

A complex type that contains a Message and a Timestamp value for changes

CidrAuthorizationContext

Provides authorization for Amazon to bring a specific IP address range to a specific AWS

CustomRoutingAccelerator

Attributes of a custom routing accelerator.

CustomRoutingAcceleratorAttributes

Attributes of a custom routing accelerator.

CustomRoutingDestinationConfiguration

For a custom routing accelerator, sets the port range and protocol for all endpoints (virtual

CustomRoutingDestinationDescription

For a custom routing accelerator, describes the port range and protocol for all endpoints

CustomRoutingEndpointConfiguration

The list of endpoint objects. For custom routing, this is a list of virtual private cloud (VPC) subnet IDs.

CustomRoutingEndpointDescription

A complex type for an endpoint for a custom routing accelerator. Each endpoint group can include one or more endpoints,

CustomRoutingEndpointGroup

A complex type for the endpoint group for a custom routing accelerator. An AWS Region can have only one endpoint group for a specific listener.

CustomRoutingListener

A complex type for a listener for a custom routing accelerator.

DestinationPortMapping

The port mappings for a specified endpoint IP address (destination).

EndpointConfiguration

A complex type for endpoints. A resource must be valid and active when you add it as an endpoint.

EndpointDescription

A complex type for an endpoint. Each endpoint group can include one or more endpoints, such as load

EndpointGroup

A complex type for the endpoint group. An AWS Region can have only one endpoint group for a specific listener.

IpSet

A complex type for the set of IP addresses for an accelerator.

Listener

A complex type for a listener.

PortMapping

Returns the ports and associated IP addresses and ports of Amazon EC2 instances in your virtual

PortOverride

Override specific listener ports used to route traffic to endpoints that are part of an endpoint group.

PortRange

A complex type for a range of ports for a listener.

SocketAddress

An IP address/port combination.

Tag

A complex type that contains a Tag key and Tag value.

Enums

AcceleratorStatus

ByoipCidrState

ClientAffinity

CustomRoutingAcceleratorStatus

CustomRoutingDestinationTrafficState

CustomRoutingProtocol

HealthCheckProtocol

HealthState

IpAddressType

Protocol