Module aws_sdk_globalaccelerator::operation

Structs

AddCustomRoutingEndpoints

Associate a virtual private cloud (VPC) subnet endpoint with your custom routing accelerator.

AdvertiseByoipCidr

Advertises an IPv4 address range that is provisioned for use with your AWS resources

AllowCustomRoutingTraffic

Specify the Amazon EC2 instance (destination) IP addresses and ports for a VPC subnet endpoint that can receive traffic

CreateAccelerator

Create an accelerator. An accelerator includes one or more listeners that process inbound connections and direct traffic

CreateCustomRoutingAccelerator

Create a custom routing accelerator. A custom routing accelerator directs traffic to one of possibly thousands

CreateCustomRoutingEndpointGroup

Create an endpoint group for the specified listener for a custom routing accelerator.

CreateCustomRoutingListener

Create a listener to process inbound connections from clients to a custom routing accelerator.

CreateEndpointGroup

Create an endpoint group for the specified listener. An endpoint group is a collection of endpoints in one AWS

CreateListener

Create a listener to process inbound connections from clients to an accelerator. Connections arrive to assigned static

DeleteAccelerator

Delete an accelerator. Before you can delete an accelerator, you must disable it and remove all dependent resources

DeleteCustomRoutingAccelerator

Delete a custom routing accelerator. Before you can delete an accelerator, you must disable it and remove all dependent resources

DeleteCustomRoutingEndpointGroup

Delete an endpoint group from a listener for a custom routing accelerator.

DeleteCustomRoutingListener

Delete a listener for a custom routing accelerator.

DeleteEndpointGroup

Delete an endpoint group from a listener.

DeleteListener

Delete a listener from an accelerator.

DenyCustomRoutingTraffic

Specify the Amazon EC2 instance (destination) IP addresses and ports for a VPC subnet endpoint that cannot receive traffic

DeprovisionByoipCidr

Releases the specified address range that you provisioned to use with your AWS resources

DescribeAccelerator

Describe an accelerator.

DescribeAcceleratorAttributes

Describe the attributes of an accelerator.

DescribeCustomRoutingAccelerator

Describe a custom routing accelerator.

DescribeCustomRoutingAcceleratorAttributes

Describe the attributes of a custom routing accelerator.

DescribeCustomRoutingEndpointGroup

Describe an endpoint group for a custom routing accelerator.

DescribeCustomRoutingListener

The description of a listener for a custom routing accelerator.

DescribeEndpointGroup

Describe an endpoint group.

DescribeListener

Describe a listener.

ListAccelerators

List the accelerators for an AWS account.

ListByoipCidrs

Lists the IP address ranges that were specified in calls to ProvisionByoipCidr, including

ListCustomRoutingAccelerators

List the custom routing accelerators for an AWS account.

ListCustomRoutingEndpointGroups

List the endpoint groups that are associated with a listener for a custom routing accelerator.

ListCustomRoutingListeners

List the listeners for a custom routing accelerator.

ListCustomRoutingPortMappings

Provides a complete mapping from the public accelerator IP address and port to destination EC2 instance

ListCustomRoutingPortMappingsByDestination

List the port mappings for a specific EC2 instance (destination) in a VPC subnet endpoint. The

ListEndpointGroups

List the endpoint groups that are associated with a listener.

ListListeners

List the listeners for an accelerator.

ListTagsForResource

List all tags for an accelerator.

ProvisionByoipCidr

Provisions an IP address range to use with your AWS resources through bring your own IP

RemoveCustomRoutingEndpoints

Remove endpoints from a custom routing accelerator.

TagResource

Add tags to an accelerator resource.

UntagResource

Remove tags from a Global Accelerator resource. When you specify a tag key, the action removes both that key and its associated value.

UpdateAccelerator

Update an accelerator.

UpdateAcceleratorAttributes

Update the attributes for an accelerator.

UpdateCustomRoutingAccelerator

Update a custom routing accelerator.

UpdateCustomRoutingAcceleratorAttributes

Update the attributes for a custom routing accelerator.

UpdateCustomRoutingListener

Update a listener for a custom routing accelerator.

UpdateEndpointGroup

Update an endpoint group. A resource must be valid and active when you add it as an endpoint.

UpdateListener

Update a listener.

WithdrawByoipCidr

Stops advertising an address range that is provisioned as an address pool.